Use allow-list for whether to insert NFC polling frames. (#6)
It turns out that if this feature is used and the device doesn't
support it then Android's NfcAdapter.enableReaderMode() silently
fails. I've asked for NfcAdapter.isReaderModeAnnotationSupported() to
be available as public API (it's currently behind a flag), until this
is done just use a white-list of devices where we know this works.

Also add UI to Developer Settings for whether to insert polling
frames, if supported.

Test: Manually tested.

Signed-off-by: David Zeuthen <zeuthen@google.com>